+++ rtvi-client-js/src/events.ts
@@ -8,7 +8,6 @@ import {
   StorageItemStoredData,
   TranscriptData,
   TTSTextData,
-  UserLLMTextData
 } from "./messages";
 import { Participant, TransportState } from "./transport";
 
@@ -48,10 +47,9 @@ export enum RTVIEvent {
   Metrics = "metrics",
 
   UserTranscript = "userTranscript",
-  UserText = "userText",
   BotTranscript = "botTranscript",
 
-  BotText = "botText",
+  BotLlmText = "botLlmText",
   BotLlmStarted = "botLlmStarted",
   BotLlmStopped = "botLlmStopped",
 
@@ -76,10 +74,10 @@ export type RTVIEvents = Partial<{
   configDescribe: (configDescription: unknown) => void;
   actionsAvailable: (actions: unknown) => void;
 
-  participantConnected: (p: Participant) => void;
-  participantLeft: (p: Participant) => void;
-  trackStarted: (track: MediaStreamTrack, p?: Participant) => void;
-  trackStopped: (track: MediaStreamTrack, p?: Participant) => void;
+  participantConnected: (participant: Participant) => void;
+  participantLeft: (participant: Participant) => void;
+  trackStarted: (track: MediaStreamTrack, participant?: Participant) => void;
+  trackStopped: (track: MediaStreamTrack, participant?: Participant) => void;
 
   availableCamsUpdated: (cams: MediaDeviceInfo[]) => void;
   availableMicsUpdated: (cams: MediaDeviceInfo[]) => void;
@@ -87,10 +85,10 @@ export type RTVIEvents = Partial<{
   micUpdated: (cam: MediaDeviceInfo) => void;
 
   botReady: (botData: BotReadyData) => void;
-  botConnected: (p: Participant) => void;
-  botDisconnected: (p: Participant) => void;
-  botStartedSpeaking: (p: Participant) => void;
-  botStoppedSpeaking: (p: Participant) => void;
+  botConnected: (participant: Participant) => void;
+  botDisconnected: (participant: Participant) => void;
+  botStartedSpeaking: (participant: Participant) => void;
+  botStoppedSpeaking: (participant: Participant) => void;
   remoteAudioLevel: (level: number, p: Participant) => void;
 
   userStartedSpeaking: () => void;
@@ -100,16 +98,14 @@ export type RTVIEvents = Partial<{
   metrics: (data: PipecatMetricsData) => void;
 
   userTranscript: (data: TranscriptData) => void;
-  userText: (text: UserLLMTextData) => void;
-  botTranscript: (data: TranscriptData) => void;
-
-  botText: (text: BotLLMTextData) => void;
-  botLlmStarted: (p: Participant) => void;
-  botLlmStopped: (p: Participant) => void;
+  botTranscript: (text: BotLLMTextData) => void;
+  botLlmText: (text: BotLLMTextData) => void;
+  botLlmStarted: () => void;
+  botLlmStopped: () => void;
 
   botTtsText: (text: TTSTextData) => void;
-  botTtsStarted: (p: Participant) => void;
-  botTtsStopped: (p: Participant) => void;
+  botTtsStarted: () => void;
+  botTtsStopped: () => void;
 
   error: (message: RTVIMessage) => void;
   messageError: (message: RTVIMessage) => void;